add feature AdminCheckAds, update Time format

+++ backend/controllers/AdvertController.js
@@ -4,8 +4,10 @@ const { Schema } = mongoose;
 const { insertData, getData, updateData, deleteData, getDataById, getNextDataId } = require('../database/Database.js');
 var str_collection = "Advert";
 
+const { update_Ads } = require('./productController.js')
+
 function formatDate(date) {
-  const formattedDate = moment(date).utcOffset('+07:00').format('DD-MM-YYYY HH:mm:ss');
+  const formattedDate = moment(date).utcOffset('+07:00').format('YYYY/MM/DD HH:mm:ss');
   return formattedDate;
 }
 
@@ -54,7 +56,16 @@ async function updateAdvert(req, res) {
     const { id } = req.params;
     const newData = req.body;
 
-    await updateData(id, newData, DataModel);
+    const product_ID_find = await DataModel.findOne({ ID: id }).exec(); // หา P_ID ด้วย ID ของ Ads
+    const product_ID = product_ID_find.P_ID;
+    if (product_ID) {
+      await updateData(id, newData, DataModel);
+      // ให้ทำการอัพเดตข้อมูลในส่วนของสถานะโฆษณาตัวของประกาศขาย
+      await update_Ads(product_ID, true);
+    }else{
+      res.status(500).json({ error: 'รหัสข้อผิดพลาด'+'Ads_ID = '+id+' P_ID = '+product_ID });
+    }
+
 
     console.log('Advert updated successfully');
     res.status(200).json({ status: true, message: 'Advert updated successfully' });
@@ -68,7 +79,15 @@ async function deleteAdvert(req, res) {
   try {
     const { id } = req.params;
 
-    await deleteData(id, DataModel);
+    const product_ID_find = await DataModel.findOne({ ID: id }).exec(); // หา P_ID ด้วย ID ของ Ads
+    const product_ID = product_ID_find.P_ID;
+    if (product_ID) {
+      await deleteData(id, DataModel);
+      // ให้ทำการอัพเดตข้อมูลในส่วนของสถานะโฆษณาตัวของประกาศขาย
+      await update_Ads(product_ID, false);
+    }else{
+      res.status(500).json({ error: 'รหัสข้อผิดพลาด'+'Ads_ID = '+id+' P_ID = '+product_ID });
+    }
 
     console.log('Advert deleted successfully');
     res.status(200).json({ status: true, message: 'Advert deleted successfully' });
@@ -113,3 +132,19 @@ async function getAdvertByProduct(req, res) {
 }
 
 module.exports = { addAdvert, listAdverts, updateAdvert, deleteAdvert, getAdvertById, getAdvertByProduct };
+
+
+
+
+
+// // ฟังก์ชันนี้จะตรวจสอบ วันหมดอายุของ การลงโฆษณาประกาศขาย และอัพเดตหากหมดอายุแล้ว
+// async function Check_Ads_Product() {
+
+//   const id = 4;
+
+//   const product_ID = await DataModel.findOne({ ID: id }, { P_ID: 1, _id: 0 }).exec();
+//   console.log('Ads_ID = '+ id +'------->',product_ID.P_ID);
+
+// }
+
+// setInterval(Check_Ads_Product, 1000);
